Move to using extern/dolphin (#1559)
* Fix pragma to use "" vs. '' as it is interpreted differently

* Attempt to fixup Padclamp

* Remove AmcExi2Stubs, add __OSTimeToSystemTime to OSTime

* Mark dsp as matched

* Mark PPCArch as extern'd

* Mark db as matched by extern

* Mark ai and ar as matched by extern

* Update dolphin/types.h with our changes to AT_ADDRESS

* Mark card as non-matching for now

* Mark axfx as extern, bring axfc.c over since it was not in this SDK version

* Revert most changes to axfx.h

* Remove duplicate function declaration

* Partial adjustments to make ax match

* Move pending code out to a directory to avoid it conflicting

* Update GX to NonMatching till manually confirmed, match GXFifo

* Remove GXFifo and Init todos

* Mark GXStubs as matching

* Mark dvd as non-matching for now

* Bring over OSReboot

* Remove ifdef

* Update include in controller.h

* Move dvderror into extern

* Fixup some of GX without major struct changes

* Match dvdqueue

* Move ARRAY_SIZE out of platform and into dolphin/types.h

* Match dvdfs and fstload

* Match hio

* Match DSPCode

* Mark AXSPB as matched

* Match AXAux

* Mark AXAlloc as matching

* Rename files to match the extern versions

* Various header fixup

* Fix tev match

* Initial resolution of linking issues and symbol name fixes

* MTX cleanup to add placeholder protos for functions that weren't in the SDK version matched originally

* Resolve match issues in GXLight and __start.c

* Fixup __start.c issues

* Remove part of TODOs

* Disable illegal pragma warnings

* Update `compile_flags.txt` to include dolphin

* Rename `__GXContexts` to `gx`

* Resolve more link issues, mark OSMemory as NonMatching

* Fixup more mtx link usages

* Fix function name

* Fixup remaining linking issues

* Fix .init section ordering

* Mark hio as NonMatching

* Fixup sysdolphin mismatches

* Mark some files as NonMatching that were causing data issues

* Fix ninja build

* Fix hio match

* Fix some matching file regressions vs master

* Undo Padclamp -> PadClamp renaming

* Add THPDec stub

* Revert "Undo Padclamp -> PadClamp renaming"

This reverts commit 3edeb3cde4a85dd1b594e017e0e80c5f932785ac.

The ninja diff github action reruns with a dirty build directory, so the
change in filename capitalization breaks the build - some part of the
build system must be case-insensitive so the object file is not created
properly.

* Also rename Pad -> pad, to match master branch naming

* Add PSMTXTrans define to mtx.h

* Declare dolphin mtx44 functions

* Fix some dolphin mtx matches

* Fix more dolphin card matches

* Remove unused src/dolphin/axfx file

* Fix melee-issues

* Fix erroneous include

* Fix dolphin card matches

* Run `ninja apply`

* Match axfx/chorus.c

* Match axfx/reverb_hi.c

* Match axfx/reverb_std.c

* Match axfx/delay.c, remove todo/axfx

* Remove card todos and axfx asm

* Match ax/axcl.c

* Match AXVPB

* Mark AXSPB as matching now that it no longer incorrectly shifts

* Run ninja apply

* Match AXOut

* Remove overlap from invalid symbol

* Match UpdateOrigin in pad.c

* Match PADOriginCallback and PADOriginUpdateCallback

* Match PADProbeCallback

* Match PADReset

* Remove TODOs

* Match OSInterrupt

* Mark OSThread as matched

* Finish matching OSAlarm and OSAudioSystem as matching

* Match OSError

* Remove redefinition of TRUNC

---------

+++ src/sysdolphin/baselib/pobj.c
@@ -1,5 +1,4 @@
 #include "forward.h"
-#include <dolphin/gx/forward.h>
 
 #include "pobj.h"
 
@@ -18,12 +17,7 @@
 
 #include <__mem.h>
 #include <math.h> // IWYU pragma: keep
-#include <dolphin/gx/GXAttr.h>
-#include <dolphin/gx/GXDisplayList.h>
-#include <dolphin/gx/GXEnum.h>
-#include <dolphin/gx/GXGeometry.h>
-#include <dolphin/gx/GXTransform.h>
-#include <dolphin/gx/GXVert.h>
+#include <dolphin/gx.h>
 #include <dolphin/mtx.h>
 #include <dolphin/os.h>
 
@@ -452,7 +446,7 @@ static void setupArrayDesc(HSD_VtxDescList* desc_list)
     if (prev_vtxdesclist_array != desc_list) {
         for (desc = desc_list; desc->attr != GX_VA_NULL; desc++) {
             if (desc->attr_type != GX_DIRECT) {
-                GXSetArray(desc->attr, (s32) desc->vertex, desc->stride);
+                GXSetArray(desc->attr, desc->vertex, desc->stride);
             }
         }
         prev_vtxdesclist_array = desc_list;
@@ -499,7 +493,7 @@ static void setupShapeAnimArrayDesc(HSD_VtxDescList* desc_list)
             case GX_VA_NBT:
                 break;
             default:
-                GXSetArray(desc->attr, (s32) desc->vertex, desc->stride);
+                GXSetArray(desc->attr, desc->vertex, desc->stride);
             }
         }
     }
@@ -1219,7 +1213,7 @@ static void PObjDispSimplePrimitive(HSD_PObj* pobj, u32 rendermode)
     setupArrayDesc(pobj->verts);
     setupVtxDesc(pobj);
 
-    GXCallDisplayList((s32) pobj->display, pobj->n_display << 5);
+    GXCallDisplayList(pobj->display, pobj->n_display << 5);
 }
 
 static void PObjDispShapeAnim(HSD_PObj* pobj, u32 rendermode)
